# packages/schemas/openapi/v0/platform.yaml
openapi: 3.1.0
info:
  title: Horus Flow API — plataforma (contrato v0)
  version: 0.1.0
  description: |
    Rutas `/platform/*` con token de ámbito `platform` (docs/api.md §2.8). Un usuario de plataforma no ve
    datos de negocio de los ISP (clientes, tráfico, hallazgos). Módulos dueños: `auth` (tenants, usuarios),
    `gateway` (overview, modo de acceso de la instalación — D19), `detection` (fuentes de reputación — D20), `flows` (exportadores no registrados), `wireguard` (hubs e IPAM), `analytics`
    (destinos remotos y copias en lectura; rol jobs).
  license:
    name: Propietaria
    identifier: LicenseRef-Horus-Proprietary
servers:
  - url: https://{domain}/api/v1
    variables:
      domain:
        default: horus.example.net
...
paths:
...
  /platform/events:
    get:
      tags: [platform]
      operationId: platformListEvents
      summary: Registro de eventos de plataforma (arranques, caídas, migraciones, degradaciones; D23)
      description: |
        Eventos de la instalación, más recientes primero: arranque y parada de cada proceso y rol con su versión,
        paradas no limpias detectadas al arrancar (`unclean_shutdown`), dependencias caídas o recuperadas según el
        healthcheck de cada rol, migraciones aplicadas, búfer de telemetría por encima del 70 %, spool activo,
        reinicios del agente WireGuard, fallos definitivos de envío de alertas y cambios de la configuración
        efectiva entre arranques. Sin datos de clientes; `tenant_id` solo en eventos de un ISP (p. ej.
        `alert_delivery_failed`). Retención `HORUS_PLATFORM_EVENTS_RETENTION` (90 días). La UI de la consola de
        plataforma queda pendiente.
      x-module: gateway
      x-scope: platform
      x-permission: platform.status.read
      x-increment: I1
      security:
        - platformBearer: []
      parameters:
        - $ref: './common.yaml#/components/parameters/Limit'
        - $ref: './common.yaml#/components/parameters/Cursor'
        - name: kind
          in: query
          description: Tipos separados por comas.
          schema: { type: string, examples: ['unclean_shutdown,dependency_down'] }
        - name: severity
          in: query
          description: Severidad mínima.
          schema: { type: string, enum: [info, warn, error] }
        - name: process
          in: query
          schema: { type: string, examples: [horus-app] }
        - name: role
          in: query
          schema: { type: string, examples: [alerts] }
        - name: since
          in: query
          schema: { $ref: './common.yaml#/components/schemas/Timestamp' }
        - name: until
          in: query
          schema: { $ref: './common.yaml#/components/schemas/Timestamp' }
      responses:
        '200':
          description: Página de eventos.
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PlatformEvent' }
                  page: { $ref: './common.yaml#/components/schemas/PageInfo' }
        '400': { $ref: './common.yaml#/components/responses/BadRequest' }
        '401': { $ref: './common.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './common.yaml#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    platformBearer: { $ref: './common.yaml#/components/securitySchemes/platformBearer' }
...
  schemas:
...
    PlatformEvent:
      type: object
      required: [id, occurred_at, kind, severity, process, instance, message]
      properties:
        id: { type: string, format: uuid }
        occurred_at: { $ref: './common.yaml#/components/schemas/Timestamp' }
        kind:
          type: string
          enum: [process_started, process_stopped, unclean_shutdown, role_started, role_stopped, role_failed, dependency_down,
            dependency_recovered, migration_applied, buffer_high, buffer_recovered, spool_active, spool_drained,
            alert_delivery_failed, config_changed, agent_restarted, container_restarted]
        severity: { type: string, enum: [info, warn, error] }
        process: { type: string, examples: [horus-app] }
        instance: { type: string, description: Hostname del contenedor (o HORUS_INSTANCE). }
        role: { type: string }
        version: { type: string }
        tenant_id: { type: [string, 'null'], format: uuid }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
          description: Datos del evento sin secretos ni datos de clientes (p. ej. dependency, error, schema, changed_keys).
        trace_id: { type: string }
...